pyproject.toml
[tool.poetry]
name = "sql-rag-agent"
version = "0.1.0"
description = "gen ai for risk identification"
authors = ["NIAMBA BEUGRE OKESS <beugre.niamba@centrale-casablanca.ma>"]
packages = [
    { include = "*", from = "risk_engine" }
]

[tool.poetry.dependencies]
# La version de Python doit être définie ici
python = ">=3.11,<4.0"

# Syntaxe correcte pour Poetry : package = "version"
pandas = "^2.2.0"
plotly = "^5.18.0"
streamlit = "^1.30.0"
python-dotenv = "^1.0.0"
llama-index = "0.12.5"
llama-cloud-services = "0.6.90"
langchain = "^1.2.7"
langchain-openai = "^1.1.7"
langgraph = "^1.0.7"
langchain-anthropic = "^1.3.1"
rank-bm25 = "^0.2.2"
matplotlib = "^3.10.8"
langchain-mistralai = "^1.1.1"
pytest = "^9.0.2"
markdown = "^3.10.2"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"